- Refactor and decouple commons from config/context etc - Remove redundant samples, code - Move wrongly merged "common/*" code to "commons/"
 tendrl-bug-id: Tendrl/commons#90
